requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
pdfkit==1.0.0
protobuf==3.20.3
python-dotenv==0.21.1
python-telegram-bot==20.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tzdata==2022.7
tzlocal==4.2
wkhtmltopdf==0.2